NEB-499: Created CardTextView interface
There are now two sub-interfaces of CardView - CardDateView and CardTextView.
The rational for this is that the implementer of CardDateView also had to
implement the method "getInsertedText" which is only used by non-composite
text view implementations (PANLayout and CardCVVText). This method has
now been moved from the super-interface into the new CardTextView interface.