NEB-543: Added a test for verifying local card config against schema
This will be run as a unit test in the sample app module, where we use
the local card configuration file in the project for stubbing
in the Mockserver class.

Validating against the remote card configuration schema will allow us to
be sure that the card configuration is always correct for the purposes of
running against our UI tests.