US1429950: fix an issue where pressing and maintaining the backspace key does not delete the entirety of the card number field (only occurring when pan formatting is enabled)
Using EditText.setText() to reformat the card number field when pan formatting is enabled is the root cause of the issue where end users are not able to delete as many characters as they want by pressing and maintaining the backspace key. Not sure why but this is interrupting the keyboard events and users are able to delete characters only up to the point where this method is called. The solution to this issue to set the text in the card number field without using setText() is to use EditText.editableText, clear the field by calling EditText.editableText.clear() and then call EditText.editableText.append(..) to set the text in the field. In addition to this, the inputType needs to be set to `datetime` to allow spaces in the card number field because the `number` inputType does not allow spaces. A side effect of this fix is that when pan formatting is enabled then the virtual keyboard displayed on the device is a datetime keyboard instead of a number keyboard; they are both very similar so this should not be an issue.